% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyTNBC.R
\name{classifyTNBC}
\alias{classifyTNBC}
\title{TNBC classification}
\usage{
classifyTNBC(mat)
}
\arguments{
\item{mat}{A matrix of gene expression with genes in rows and samples in columns (rownames corresopnding to gene symbol).}
}
\value{
A vector of assigned subtypes.
}
\description{
Assign the TNBC consensus molecular subtype to TNBC samples.
}
