% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDS.R
\name{computeDS}
\alias{computeDS}
\title{Computation of drug signature scores.}
\usage{
computeDS(exp.mat, pred)
}
\arguments{
\item{exp.mat}{A matrix of gene expression with genes in rows and samples in columns (rownames corresopnding to gene symbol).}

\item{pred}{A vector of predicted consensus molecular subtypes.}
}
\value{
A matrix of drug signature scores.
}
\description{
Computes drug signature scores. Also draws heatmap representing the average signature scores for each subtype.
}
\details{
Drug signature scores are the average of expression values of genes included in gene sets from MSigDB.
}
\references{
Liberzon,A. et al. (2011) Molecular signatures database (MSigDB) 3.0. \emph{Bioinformatics},27,1739-40.
}
