% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGES.R
\name{computeESTIMATEscore}
\alias{computeESTIMATEscore}
\title{Computation of stromal and immune scores}
\usage{
computeESTIMATEscore(mat)
}
\arguments{
\item{mat}{A matrix of gene expression with genes in rows and samples in columns (rownames corresopnding to gene symbol).}
}
\value{
A data frame containing stromal and immune scores
}
\description{
Computes stromal and immune scores. This function was borrowed from the \code{estimate} package and modified to accept R object as input.
}
\keyword{internal}
