% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGES.R
\name{computeGES}
\alias{computeGES}
\title{Computation of gene expression signature scores.}
\usage{
computeGES(expr, pred, rnaseq = FALSE)
}
\arguments{
\item{expr}{A \code{SummarizedExperiment} object or a matrix containig gene
expression profiles. If input is a \code{SummarizedExperiment}, the first
element in the assays list should be a matrix of gene expression.
Rows and columns of the gene expression matrix correspond to genes and
samples, respectively (rownames must be to gene symbols).}

\item{pred}{A vector of predicted consensus molecular subtypes.}

\item{rnaseq}{logical to determine if input data is
RNA-Seq gene expression profile. By default, it is FALSE.}
}
\value{
A matrix of gene expression signature scores.
}
\description{
Computes gene expression signature scores. Also draws boxplots
representing the average signature scores for each subtype.
}
\details{
\code{computeGES} calculates the following 7 gene expression
signature scores:
\itemize{
\item EMT (epithelial-mesenchymal transition):
average of expression values of genes included in the
EMT signature published by \cite{Tan et al. (2014)}.
\item Stromal: stromal score representing the presence of
stromal cells in tumor tissues (computed using the ESTIMATE algorithm).
\item Immune: immune score representing the presence of
immune cells in tumor tissues (computed using the ESTIMATE algorithm).
\item Microenvironment: microenvironment score representing the sum
of all immune and stromal cell types (computed using xCell)
\item Stemness: stemness index computed using the method developed by
\cite{Malta et al. (2018)}.
\item Hormone: average of expression values of AR, ERBB2, ESR1, and PGR.
\item CIN (chromosomal instability): average of expression values of genes
included in the CIN70 signature published by \cite{Carter et al. (2006)}.
}
}
\examples{
# Load gene expression profiles of TNBC samples
data(GSE25055)

# Predict consensus molecular subtypes of TNBC samples
prediction <- predictCMS(expr = GSE25055)

# Compute gene expression signature scores
resultGES <- computeGES(expr = GSE25055, pred = prediction, rnaseq = FALSE)
}
\references{
Aran, D. et al. (2017). xCell: digitally portraying
the tissue cellular heterogeneity landscape. \emph{Genome biology}, 18, 220.

Carter, S.L. et al. (2006). A signature of chromosomal
instability inferred from gene expression profiles predicts
clinical outcome in multiple human cancers. \emph{Nature genetics}, 38, 1043.

Malta, T.M. et al. (2018). Machine learning
identifies stemness features associated with
oncogenic dedifferentiation. \emph{Cell}, 173, 338-354.

Tan, T.Z. et al. (2014). Epithelial-mesenchymal
transition spectrum quantification and its efficacy in deciphering
survival and drug responses of cancer
patients. \emph{EMBO molecular medicine}, 6, 1279-93.

Yoshihara, K. et al. (2013). Inferring
tumour purity and stromal and immune cell admixture
from expression data. \emph{Nature communications}, 4, 2612.
}
