\name{GSE25055.clinical}
\alias{GSE25055.clinical}
\docType{data}
\title{
TNBC clinical annotation data
}
\description{
This dataset contains disease-free survival (DFS) of TNBC samples
from GSE25055. Clinical annotations were obtained from
the \code{curatedBreastData} package.
}
\source{
https://bioconductor.org/packages/release/data/experiment/html/curatedBreastData.html
}
\references{
Hatzis, D. et al. (2017). xCell: digitally portraying the tissue
cellular heterogeneity landscape. \emph{Genome biology}, 18, 220.
}
\examples{
data(GSE25055.clinical)
}
\keyword{datasets}
