\name{GSE25055.exprs}
\alias{GSE25055.exprs}
\docType{data}
\title{
TNBC gene expression data
}
\description{
This is a gene expression dataset of TNBC samples from GSE25055.
We obtained gene expression profiles of entire breast cancer
samples from the \code{curatedBreastData} package and extracted
TNBC samples based on the expression profiles and
immunohistochemistry results.
}
\source{
https://bioconductor.org/packages/release/data/experiment/html/curatedBreastData.html
}
\references{
Hatzis, C. et al. (2011). A genomic predictor of response and survival following 
taxane-anthracycline chemotherapy for invasive breast cancer \emph{JAMA}, 305, 1873-81.
}
\examples{
data(GSE25055.exprs)
}
\keyword{datasets}
