% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCMS.R
\name{performCAMERA}
\alias{performCAMERA}
\title{Gene set analysis using CAMERA}
\usage{
performCAMERA(mat, predicted)
}
\arguments{
\item{mat}{A log-transformed matrix of gene expression}

\item{predicted}{A vector of predicted subtypes}
}
\value{
A matrix with adjusted p-value for each subtype
}
\description{
Performs a competitive gene set test (CAMERA) for each subtype. This function is borrowed from the \code{CMScaller} package and modified to test for TNBC consensus molecular subtype.
}
