% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performGSVA.R
\name{performGSVA}
\alias{performGSVA}
\title{Gene set variation analysis}
\usage{
performGSVA(exp.mat, pred)
}
\arguments{
\item{exp.mat}{A matrix of gene expression with genes in rows and samples in columns (rownames corresopnding to gene symbol).}

\item{pred}{A vector of predicted consensus molecular subtypes.}
}
\value{
A matrix of GSVA enrichment scores.
}
\description{
Performs GSVA on the hallmark gene sets. Also draws a heatmap representing GSVA scores.
}
\references{
Liberzon,A. et al. (2015) The molecular signatures database hallmark gene set collection. \emph{Cell systems},1,417-425.
}
