% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHR.R
\name{plotHR}
\alias{plotHR}
\title{Forest plot of hazard ratios}
\usage{
plotHR(exp.mat, gene.symbol, pred, time, event, by.subtype = TRUE)
}
\arguments{
\item{exp.mat}{A matrix of gene expression with genes in rows and samples in columns (rownames corresopnding to gene symbol).}

\item{gene.symbol}{A vector of gene symbols for which hazard ratios are computed.}

\item{pred}{A vector of predicted consensus molecular subtypes.}

\item{time}{A vector of the follow-up time.}

\item{event}{A vector representing survival status (0 = alive, 1 = dead).}

\item{by.subtype}{A logical to determine if subtype-specific hazard ratios are computed (default is TRUE).}
}
\description{
Produces a forest plot of hazard ratios for each gene. Also draws a forest plot of subtype-specific hazard ratios.
}
