% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDS.R
\name{computeDS}
\alias{computeDS}
\title{Computation of drug signature scores.}
\usage{
computeDS(expr, pred, gene.set = NULL)
}
\arguments{
\item{expr}{A \code{SummarizedExperiment} object or a matrix containig gene
expression profiles. If input is a \code{SummarizedExperiment}, the first
element in the assays list should be a matrix of gene expression.
Rows and columns of the gene expression matrix correspond to genes and
samples, respectively (rownames must be to gene symbols).}

\item{pred}{A vector of predicted consensus molecular subtypes.}

\item{gene.set}{A user-provided list of gene sets associated with
drug response. Names of gene sets must follow the format of
[DRUG NAME]_[RESISTANCE/RESPONSE]_[UP/DN] (e.g. CISPLATIN_RESISTANCE_DN).}
}
\value{
A matrix of drug signature scores.
}
\description{
Computes drug signature scores. Also draws heatmap representing
the average signature scores for each subtype.
}
\details{
Drug signature scores are the average of expression values of
genes included in gene sets from MSigDB.
}
\examples{
# Load gene expression profiles of TNBC samples
data(GSE25055)

# Predict consensus molecular subtypes of TNBC samples
prediction <- predictCMS(expr = GSE25055)

# Compute drug signature scores
resultDS <- computeDS(expr = GSE25055, pred = prediction)
}
\references{
Liberzon, A. et al. (2011). Molecular signatures database
(MSigDB) 3.0. \emph{Bioinformatics}, 27, 1739-40.
}
