\name{GSE25055.clinical}
\alias{GSE25055.clinical}
\docType{data}
\title{
TNBC clinical annotation data
}
\description{
This dataset contains disease-free survival (DFS) of TNBC samples
from GSE25055. Clinical annotations were obtained from
the \code{curatedBreastData} package.
}
\source{
https://bioconductor.org/packages/release/data/experiment/html/curatedBreastData.html
}
\references{
Curtis, C. et al. (2012). The genomic and transcriptomic architecture of 2,000 breast tumours
reveals novel subgroups. \emph{Nature}, 486, 346.
}
\examples{
data(GSE25055.clinical)
}
\keyword{datasets}
