\name{GSE25055.clinical}
\alias{GSE25055.clinical}
\docType{data}
\title{
TNBC clinical annotation data
}
\description{
This dataset contains disease-free survival (DFS) of TNBC samples
from GSE25055. Clinical annotations were obtained from
the \code{curatedBreastData} package.
}
\source{
https://bioconductor.org/packages/release/data/experiment/html/curatedBreastData.html
}
\references{
Hatzis, C. et al. (2011). A genomic predictor of response and survival following 
taxane-anthracycline chemotherapy for invasive breast cancer \emph{JAMA}, 305, 1873-81.
}
\examples{
data(GSE25055.clinical)
}
\keyword{datasets}
