\name{GSE25055.exprs}
\alias{GSE25055.exprs}
\docType{data}
\title{
TNBC gene expression data
}
\description{
This is a gene expression dataset of TNBC samples from GSE25055.
We obtained gene expression profiles of entire breast cancer
samples from the \code{curatedBreastData} package and extracted
TNBC samples based on the expression profiles and
immunohistochemistry results.
}
\source{
https://bioconductor.org/packages/release/data/experiment/html/curatedBreastData.html
}
\references{
Curtis, C. et al. (2012). The genomic and transcriptomic architecture of 2,000 breast tumours
reveals novel subgroups. \emph{Nature}, 486, 346.
}
\examples{
data(GSE25055.exprs)
}
\keyword{datasets}
