\name{GSE25055.exprs}
\alias{GSE25055.exprs}
\docType{data}
\title{
TNBC gene expression data
}
\description{
This is a gene expression dataset of TNBC samples from GSE25055. 
We obtained gene expression profiles of entire breast cancer 
samples from the \code{curatedBreastData} package and extracted 
TNBC samples based on the expression profiles and 
immunohistochemistry results.
}
\source{
https://bioconductor.org/packages/release/data/experiment/
html/curatedBreastData.html
}
\references{
Hatzis,D. et al. (2017) xCell: digitally portraying the tissue 
cellular heterogeneity landscape. \emph{Genome biology},18,220.
}
\examples{
data(GSE25055.exprs)
}
\keyword{datasets}
