% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHR.R
\name{plotHR}
\alias{plotHR}
\title{Forest plot of hazard ratios}
\usage{
plotHR(exp.mat, gene.symbol, pred, time, event, by.subtype = TRUE)
}
\arguments{
\item{exp.mat}{A matrix of gene expression with genes in rows and
samples in columns (rownames corresopnding to gene symbol).}

\item{gene.symbol}{A vector of gene symbols for
which hazard ratios are computed.}

\item{pred}{A vector of predicted consensus molecular subtypes.}

\item{time}{A vector of the follow-up time.}

\item{event}{A vector representing survival status (0 = alive, 1 = dead).}

\item{by.subtype}{A logical to determine if subtype-specific
hazard ratios are computed (default is TRUE).}
}
\value{
A forest plot of hazard ratios.
}
\description{
Produces a forest plot of hazard ratios for each gene.
Also draws a forest plot of subtype-specific hazard ratios.
}
\examples{
# Load gene expression profiles and clinical information of TNBC samples
data(GSE25055.exprs)
data(GSE25055.clinical)

# Predict consensus molecular subtypes of TNBC samples
predictions <- predictCMS(exp.mat = GSE25055.exprs)

# Forest plot of hazard ratios for input genes
plotHR(exp.mat = GSE25055.exprs, gene.symbol = c("FOXC1", "PCSK5", "CD19", "NAT1"),
       pred = predictions, time = GSE25055.clinical$DFS.month,
       event = GSE25055.clinical$DFS.status, by.subtype = FALSE)

# Subtype-specific forest plot of hazard ratios for input genes
plotHR(exp.mat = GSE25055.exprs, gene.symbol = c("FOXC1", "PCSK5", "CD19", "NAT1"),
       pred = predictions, time = GSE25055.clinical$DFS.month,
       event = GSE25055.clinical$DFS.status, by.subtype = TRUE)
}
