% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKM.R
\name{plotKM}
\alias{plotKM}
\title{Subtype-specific survival curves}
\usage{
plotKM(pred, time, event)
}
\arguments{
\item{pred}{A vector of predicted consensus molecular subtypes.}

\item{time}{A vector of the follow-up time.}

\item{event}{A vector representing survival status (0 = alive, 1 = dead).}
}
\description{
Produces Kaplan-Meier survival curves for each subtype.
}
