% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCMS.R
\name{predictCMS}
\alias{predictCMS}
\title{TNBC consensus molecular subtype prediction}
\usage{
predictCMS(exp.mat)
}
\arguments{
\item{exp.mat}{A matrix of gene expression with genes in rows and samples in columns (rownames corresopnding to gene symbol).}
}
\value{
A vector of assigned subtypes.
}
\description{
This function predicts the TNBC consensus molecular subtype of TNBC samples.
}
\examples{
# Load gene expression profiles of TNBC samples
data(GSE25055.exprs)

# Predict consensus molecular subtypes of TNBC samples
predictions <- predictCMS(exp.mat = GSE25055.exprs)
table(predictions)
}
